image-labelling-object-detection-app/requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
fsspec==2025.5.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.0.206
urllib3==2.5.0
Werkzeug==3.1.3
